pyproject.toml
[tool.poetry]
name = "bgo"
version = "0.1.4"
description = "About"
authors = ["bog (before going out) is sexy terminal util to view the weather"]
readme = "README.md"
packages = [{include = "bgo"}]

[tool.poetry.scripts]
bgo = 'bgo.__main__:main'

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
rich = "^13.7.0"
requests = "^2.31.0"
types-requests = "^2.31.0.20240218"
pre-commit = "^3.6.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
ruff = "==0.0.240"
mypy = "^1.8.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
select = ["F", "E", "W", "C90",
          "I", "N", "S", "B", "A",
          "ISC", "T20", "Q", "PTH"]

ignore = ["A003"]
line-length = 88
